Ensure that PATH env var doesn't have empty components.
Running find is dangerous if the PATH env var contains either "::" or
".". Both these forms slip in quite innocuously and become an issue much
later when looking for files using find. So, the changes here are to
ensure that I do not introduce "::" unknowingly. Learned about how to
check whether path is already in env var from ~/.fzf.bash .